:recycle: inline playmarket test fixtures, drop per-cog conftest
Address review: no other cog uses a per-cog conftest — fixtures live in the
test files (satisfy, with 12 files, does the same). Move alice/bob/carol,
make_context, clock and the read helpers into market_test.py (and a small cog
fixture into market_unit_test.py) where they're coupled to the tests, and
delete tests/cogs/playmarket/conftest.py. The shared in_memory_db stays in
tests/fixtures.
